Move field list save logic to FormEntryViewModel
